New version 0.2.1
License was upgraded to GPLv3+ to to make it compatible with other projects
under Apache License 2.0 and also to explicitly make it incompatible with
existing Freedesktop projects due to misusing of their own CoC to support
license violation of other projects. See the README file for details.

This project is still open source, stay open source and be GPL compatible
for all other projects which do not violate GPL and other licenses. This
change affects only projects which violates GPL licenses either directly or
transitionally like Freedesktop projects.